Refactor session_keys to agents package
