+++ core/agents/lastfm/agent.go
@@ -7,7 +7,6 @@ import (
 	"github.com/navidrome/navidrome/conf"
 	"github.com/navidrome/navidrome/consts"
 	"github.com/navidrome/navidrome/core/agents"
-	"github.com/navidrome/navidrome/core/agents/sessionkeys"
 	"github.com/navidrome/navidrome/core/scrobbler"
 	"github.com/navidrome/navidrome/log"
 	"github.com/navidrome/navidrome/model"
@@ -21,7 +20,7 @@ const (
 
 type lastfmAgent struct {
 	ds          model.DataStore
-	sessionKeys *sessionkeys.SessionKeys
+	sessionKeys *agents.SessionKeys
 	apiKey      string
 	secret      string
 	lang        string
@@ -34,7 +33,7 @@ func lastFMConstructor(ds model.DataStore) *lastfmAgent {
 		lang:        conf.Server.LastFM.Language,
 		apiKey:      conf.Server.LastFM.ApiKey,
 		secret:      conf.Server.LastFM.Secret,
-		sessionKeys: &sessionkeys.SessionKeys{DataStore: ds, KeyName: sessionKeyProperty},
+		sessionKeys: &agents.SessionKeys{DataStore: ds, KeyName: sessionKeyProperty},
 	}
 	hc := &http.Client{
 		Timeout: consts.DefaultHttpClientTimeOut,
