+++ core/agents/session_keys.go
@@ -1,4 +1,4 @@
-package sessionkeys
+package agents
 
 import (
 	"context"
