+++ core/agents/session_keys_test.go
@@ -1,4 +1,4 @@
-package sessionkeys
+package agents
 
 import (
 	"context"
