--- core/agents/sessionkeys/session_keys_suite_test.go
@@ -1,17 +1,0 @@
-package sessionkeys
-
-import (
-	"testing"
-
-	"github.com/navidrome/navidrome/log"
-	"github.com/navidrome/navidrome/tests"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestSessionKeys(t *testing.T) {
-	tests.Init(t, false)
-	log.SetLevel(log.LevelCritical)
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "SessionKeys Test Suite")
-}
